src/ml_model/requirements.txt
# src/ml_model/requirements.txt
google-cloud-storage
numpy
